refactor(planner): 로딩 스켈레톤 컴포넌트 경로 수정
